(fix) convert todo tasks into events; integrated todo list apis
