+++ ui/MorganizerUI/src/app/to-do-list/edit-task/edit-task.component.html
@@ -1,9 +1,9 @@
 <div [@todoInOutAnimation]>
-  <div style="border-bottom: solid #d1d1d1; border-bottom-width: 2px;">
+  <div style="border-bottom: solid #d1d1d1; border-bottom-width: 2px">
     <div class="row">
       <div class="col-md-2">
         <mat-icon
-          style="cursor: pointer; color: gray;"
+          style="cursor: pointer; color: gray"
           matRipple
           matTooltip="Back"
           [matTooltipPosition]="'right'"
@@ -13,12 +13,12 @@
           >keyboard_arrow_left</mat-icon
         >
       </div>
-      <div class="col-md-8" style="text-align: center; color: gray;">
+      <div class="col-md-8" style="text-align: center; color: gray">
         <mat-label>Edit Task</mat-label>
       </div>
       <div class="col-md-2">
         <mat-icon
-          style="float: right;"
+          style="float: right"
           matTooltip="Delete Task"
           class="material-icons md-18 delete-task"
           (click)="removeTask()"
@@ -27,8 +27,8 @@
       </div>
     </div>
   </div>
-  <div class="flex-vertical" style="margin-top: 10px; margin-left: 10px;">
-    <mat-form-field appearance="outline" style="width: 100%;">
+  <div class="flex-vertical" style="margin-top: 10px; margin-left: 10px">
+    <mat-form-field appearance="outline" style="width: 100%">
       <mat-label>Title</mat-label>
       <input
         matInput
@@ -36,7 +36,7 @@
         placeholder="Enter title"
       />
     </mat-form-field>
-    <mat-form-field appearance="outline" style="width: 100%;">
+    <mat-form-field appearance="outline" style="width: 100%">
       <mat-label>Task Details</mat-label>
       <textarea
         matInput
@@ -47,7 +47,7 @@
         placeholder="Add details"
       ></textarea>
     </mat-form-field>
-    <mat-form-field appearance="outline" style="width: 100%;">
+    <mat-form-field appearance="outline" style="width: 100%">
       <mat-label>Calendar</mat-label>
       <mat-select
         [formControl]="selectedCalendar"
@@ -60,9 +60,9 @@
       </mat-select>
     </mat-form-field>
 
-    <mat-form-field appearance="outline" style="width: 100%;">
+    <mat-form-field appearance="outline" style="width: 100%">
       <input
-        [(ngModel)]="data[taskIndex].duedate"
+        [(ngModel)]="data[taskIndex].dueDate"
         matInput
         [ngxMatDatetimePicker]="picker"
         placeholder="Due date and time"
